+++ SmartFarmer.App/Communication/IFarmerDeviceHandler.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 public interface IFarmerDeviceHandler : IFarmerPointNotifier
 {
     Task<bool> MoveOnGridAsync(FarmerPoint position, CancellationToken token);
-    Task<bool> MoveArmAtheightAsync(int heightInCm, CancellationToken token);
+    Task<bool> MoveArmAtheightAsync(double heightInCm, CancellationToken token);
     Task<bool> TurnArmToDegreesAsync(double degrees, CancellationToken token);
     Task<bool> PointDeviceAsync(double degrees, CancellationToken token);
 }
