+++ SmartFarmer.Core/Tasks/Movement/IFarmerMoveArmAtHeight.cs
@@ -6,7 +6,7 @@ namespace SmartFarmer.Tasks.Movement
 {
     public interface IFarmerMoveArmAtHeight : IFarmerTask
     {
-        Task MoveToHeight(int heightInCm, CancellationToken token);
-        int GetCurrentHeight();
+        Task MoveToHeight(double heightInCm, CancellationToken token);
+        double GetCurrentHeight();
     }
 }
